pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "flamingo-inference"
version = "0.1.0"
description = "High-availability inference server for Music Flamingo model"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "Koosha" }
]
keywords = ["music", "audio", "flamingo", "inference", "llm", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.0",
    "transformers>=4.40",
    "accelerate>=0.26",
    "fastapi>=0.109",
    "uvicorn[standard]>=0.27",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyzmq>=25.0",
    "msgspec>=0.18",
    "torchaudio>=2.0",
    "librosa>=0.10",
    "soundfile>=0.12",
    "prometheus-client>=0.19",
    "aiohttp>=3.9",
    "aiofiles>=23.0",
    "pyyaml>=6.0",
    "numpy>=1.24",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "httpx>=0.26",
    "ruff>=0.1",
    "mypy>=1.8",
]

[project.scripts]
flamingo = "flamingo_inference.entrypoints.cli.main:main"

[project.urls]
Homepage = "https://github.com/koosha/flamingo-inference"
Documentation = "https://github.com/koosha/flamingo-inference#readme"

[tool.hatch.build.targets.wheel]
packages = ["flamingo_inference"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
